+++ SmartScaleMinimalApi/SmartScale.cs
@@ -32,6 +32,6 @@
         /// <summary>
         /// 建立时间
         /// </summary>
-        public int Ctime { get; set; }
+        public long Ctime { get; set; }
     }
 }
